[mod planche] conversionde type de i32 vers i32 inutile (numero_ogn):194

+++ src/planche/mod.rs
@@ -212,7 +212,7 @@ impl MettreAJour for Planche {
             log::error!("Mise a jour impossible: les dates ne sont pas les mêmes !");
         } else if mise_a_jour.champ_mis_a_jour.clone() == "nouveau" {
             vols.push(Vol {
-                numero_ogn: mise_a_jour.numero_ogn as i32,
+                numero_ogn: mise_a_jour.numero_ogn,
                 aeronef: mise_a_jour.nouvelle_valeur.clone(),
                 code_vol: String::new(),
                 code_decollage: String::new(),
